+++ umongo/document.py
@@ -251,6 +251,9 @@ class DocumentImplementation(BaseDataObject, Implementation, metaclass=MetaDocum
         """
         return not self.is_created or self._data.is_modified()
 
+    def required_validate(self):
+        self._data.required_validate()
+
     def items(self):
         return self._data.items()