+++ HISTORY.rst
@@ -5,6 +5,7 @@ History
 dev
 ---
 
+* Required field inside embedded document are handled correctly (see #61)
 * Document support marshmallow's pre/post processors
 
 0.10.0 (2016-09-29)